requirements/local.txt
alembic==1.4.3
asn1crypto==0.24.0
click==7.1.2
cryptography==2.1.4
defusedxml==0.5.0
djangorestframework==3.12.2
enum34==1.1.6
Flask==1.1.2
Flask-API==1.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
Mako==1.1.3
MarkupSafe==1.1.1
pycairo==1.16.2
pycrypto==2.6.1
pygobject==3.26.1
pyliblzma==0.5.3
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pyxdg==0.25
SecretStorage==2.3.1
six==1.15.0
SOAPpy==0.12.22
SQLAlchemy==1.3.19
typing==3.7.4.1
virtualenv==16.7.5
Werkzeug==1.0.1
wstools==0.4.3
WTForms==2.3.3
